+++ litex/soc/software/liblitesdcard/sdcard.h
@@ -65,7 +65,7 @@ int sdcard_wait_response(void);
 /* SDCard clocker functions                                              */
 /*-----------------------------------------------------------------------*/
 
-void sdcard_set_clk_freq(uint32_t clk_freq, int show);
+void sdcard_set_clk_freq(unsigned long clk_freq, int show);
 
 /*-----------------------------------------------------------------------*/
 /* SDCard commands functions                                             */
